Split MoleculeSelector component to its own module
Closes #29
